feat: Market Price | Refactoring
