+++ app/screens/TradeScreen/ErrorContext.ts
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const ErrorContext = React.createContext({
+  errors: [],
+  setErrors: (_: any) => {},
+});
+
+export default ErrorContext;
